requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow==2.16.1
python-multipart
gunicorn